Add optional NginX authentication to C3CI Jenkins
Authentication is optional and disabled by default.
See the spec file for details on how to enable authentication.

The commit also changes the configure_jenkins job to access jenkins
on the internal port rather than through NginX.

Squashed commit of the following:

commit 44579597c38c7b0912ece1b5c6c7f0cda4c79d58
Author: Troy Astle <troy.astle@gmail.com>
Date:   Tue Oct 14 13:01:45 2014 +0100

    Still working on travis build settings

commit 16b8400a477baa8c51b4e58d16318eb33a44cfcb
Author: Troy Astle <troy.astle@gmail.com>
Date:   Tue Oct 14 12:53:28 2014 +0100

    Still working on travis build settings

commit 565b8f298adabb1bdf4e34a94b12f28920bf772a
Author: Troy Astle <troy.astle@gmail.com>
Date:   Tue Oct 14 12:42:07 2014 +0100

    Fix travis build paths & add travis link

commit 403b5597c7d7280a9a54df13807a3512615fb177
Author: Troy Astle <troy.astle@gmail.com>
Date:   Tue Oct 14 12:36:01 2014 +0100

    Add Auth at the NginX level.

    As a result move access for configure Jenkins to the internal port for Jenkins.

+++ src/configure_jenkins/lib/jenkins/tasks/update_config.rb
@@ -41,7 +41,9 @@ module Jenkins
       end
 
       def copy_maven_config_into_place
-        FileUtils.cp src_maven_config_path, existing_maven_config_path
+        if (File.exists? src_maven_config_path )
+            FileUtils.cp src_maven_config_path, existing_maven_config_path
+        end
       end
 
       def src_config_path
